PC Programming

+++ str_utils.c
@@ -1,8 +1,8 @@
 #include "str_utils.h"
 
-uint8_t strLen(const char* str)
+uint16_t strLen(const char* str)
 {
-	uint8_t len = 0;
+	uint16_t len = 0;
 	while (*(str + len) != '\0')
 	{
 		len++;
